user-service/requirements.txt
fastapi
uvicorn
python-jose
pydantic
psycopg2-binary
sqlalchemy
alembic
pytest
httpx<1.0.0
pydantic-settings